Use composer --working-dir
Similar to git -C it better to tell Composer which working directory to use
instead of actually changing the working directory.

Another small nit is that the `--dev` option is now the default.
It may be removed in the future.
Use `--no-dev` if you want only the `require` section to be used.