Make regreph a console app with two commands
`bin/regreph` is now a `symfony/console` application with two commands:
 - bench - bench a library based on a test file
 - regreph - bench the working dir and a past refspec of a library and compare the results

The regreph command is no longer using the bench command through the shell.
Instead it is using the common Regreph\Bench classes to perform the benchmarking.

---

WORK IN PROGRESS
================

The `regreph` command is still work in progress

One of thie things not working right now is the benchmarking of the refspec.
It does not find the tests.